Moved stage parsing to its own file
